+++ lib/dialogs/t_confirm_dialog.dart
@@ -1,17 +1,17 @@
 import 'package:flutter/material.dart';
 
 class TConfirmDialog extends StatelessWidget {
-  final String title;
-  final String contentText;
-  final String cancelText;
-  final String submitText;
+  final String? title;
+  final String? cancelText;
+  final String? submitText;
   final void Function()? onCancel;
   final void Function() onSubmit;
+  final String contentText;
   const TConfirmDialog({
     super.key,
-    this.title = 'Confirm',
-    this.cancelText = 'Cancel',
-    this.submitText = 'Submit',
+    this.title,
+    this.cancelText,
+    this.submitText,
     this.onCancel,
     required this.contentText,
     required this.onSubmit,
@@ -20,7 +20,7 @@ class TConfirmDialog extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return AlertDialog(
-      title: Text(title),
+      title: Text(title ?? 'Confirm'),
       content: Text(contentText),
       actions: [
         TextButton(
@@ -30,14 +30,14 @@ class TConfirmDialog extends StatelessWidget {
               onCancel!();
             }
           },
-          child: Text(cancelText),
+          child: Text(cancelText ?? 'Cancel'),
         ),
         TextButton(
           onPressed: () {
             Navigator.of(context).pop(true);
             onSubmit();
           },
-          child: Text(submitText),
+          child: Text(submitText ?? 'Confirm'),
         ),
       ],
     );